refactor: split cr-infra/video/mod.rs (1366→7 files, max 442) (closes #438)
Part of #436. Decomposes the god-module into:
- mod.rs (125) — types + thin public API dispatchers
- transcode.rs (442) — ensure_container, extract_audio, convert_for_whatsapp
- extractors/ytdlp.rs (342) — yt-dlp subprocess wrangling
- extractors/seznam.rs (258) — Seznam/Novinky direct HTTP
- extractors/nova.rs (113) — Czech-proxy Nova.cz fallback
- extractors/instagram.rs (89) — Instagram base64 extraction

Public API surface unchanged. cr-infra + cr-web compile clean.
