+++ cr-infra/src/video/extractors/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod instagram;
+pub(crate) mod nova;
+pub(crate) mod seznam;
+pub(crate) mod ytdlp;
